+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## [0.2.0] - 2025-02-24
+## [0.2.1] - 2025-03-17
+
+### Added
+- Fix using user specified template
+- Improve domain matching
+
+## [0.2.0] - 2025-03-15
 
 ### Added
 - Allow specifying lists of block patterns